fix(admin-navbar): add "use client" directive for component
The addition of the "use client" directive at the top of the AdminNavbar.jsx file ensures that this component is rendered on the client side. This change addresses potential issues with component behavior in a Next.js environment where server-side rendering could lead to unintended outcomes or errors, particularly when using hooks like usePathname and useRouter.